requirements.txt
streamlit==1.27.0
ultralytics==8.0.210
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.5.0
watchdog